refactor(invites): merge invite hooks into one useInviteFlow
Collapses useInviteValidation + useInviteAcceptance (both only used by the
root route) into a single flow hook, drops the AuthContext comment, and
reworks the test to drive the whole flow through the real query/mutation
hooks with only the supabase client mocked.
